+++ packages/kaisel/lib/src/kaisel_router_config.dart
@@ -6,6 +6,7 @@ import 'kaisel_default_page.dart';
 import 'kaisel_page_wrapper.dart';
 import 'kaisel_route_information_parser.dart';
 import 'kaisel_router_delegate.dart';
+import 'kaisel_screen_signal.dart';
 import 'kaisel_stack_restorer.dart';
 
 /// A ready-made [RouterConfig] that bundles a [KaiselRouter] and a
@@ -64,6 +65,7 @@ class KaiselRouterConfig<R extends KaiselRoute>
     bool androidPredictiveBack = false,
     KaiselModalBuilder? modalBuilder,
     KaiselObserversBuilder? observers,
+    KaiselScreenCallback? onScreenChanged,
     Listenable? reevaluateOn,
     GlobalKey<NavigatorState>? navigatorKey,
     String? restorationScopeId,
@@ -86,6 +88,7 @@ class KaiselRouterConfig<R extends KaiselRoute>
         androidPredictiveBack: androidPredictiveBack,
         modalBuilder: modalBuilder,
         observers: observers,
+        onScreenChanged: onScreenChanged,
         reevaluateOn: reevaluateOn,
         navigatorKey: navigatorKey,
         restorationScopeId: restorationScopeId,
@@ -110,6 +113,7 @@ class KaiselRouterConfig<R extends KaiselRoute>
     bool androidPredictiveBack = false,
     KaiselModalBuilder? modalBuilder,
     KaiselObserversBuilder? observers,
+    KaiselScreenCallback? onScreenChanged,
     Listenable? reevaluateOn,
     GlobalKey<NavigatorState>? navigatorKey,
     String? restorationScopeId,
@@ -132,6 +136,7 @@ class KaiselRouterConfig<R extends KaiselRoute>
         androidPredictiveBack: androidPredictiveBack,
         modalBuilder: modalBuilder,
         observers: observers,
+        onScreenChanged: onScreenChanged,
         reevaluateOn: reevaluateOn,
         navigatorKey: navigatorKey,
         restorationScopeId: restorationScopeId,